Reduce number of calls to source API in getV2ClientNamesRecord
